Update README.md
In this project, you are required to determine the shortest path between two cities using the Uniform Cost Search (UCS) algorithm. Figure 1 illustrates the road map provided for this assignment.

Figure 1. A simplified road map of some Turkish cities.

To-Do

You are going to develop a Python application, which is expected to meet the following requirements:

R1. The road map should be initialized from a csv file.

R2. The user should be able to specify the path of the road map file.

R3. The user should be able to specify the start (current city).

R4. The user should be able to specify the end (target city).